refactor: move version-specific LocalHook into _af2/_af3 top-level folders
Reorganise so the version-scoped hook code sits alongside the matching
compat layer, API and plugin classes in astronomer_starship/_af2/ and
_af3/, matching the existing pattern in those folders instead of
introducing a parallel _af2/_af3 split inside providers/starship/hooks/.

* astronomer_starship/_af2/starship_hook.py  (direct-DB LocalHook)
* astronomer_starship/_af3/starship_hook.py  (HTTP LocalHook)
* providers/starship/hooks/starship.py       (unchanged public surface,
                                              tail dispatch now imports
                                              from the top-level dirs)
